pyproject.toml
[project]
name = "apollo-io-mcp-server"
version = "0.1.0"
description = "MCP server that exposes the Apollo.io API functionalities as tools"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.6.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "vcrpy>=6.0.0",
    "pytest-recording>=0.13.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: marks tests that require API credentials (deselect with '-m \"not integration\"')",
]
# By default, skip integration tests
addopts = "-m 'not integration'"
